+++ lib/src/core/commons/widgets/dropdown.widget.dart
@@ -1,4 +1,4 @@
-import 'package:aplikasi_point_of_sale/src/core/commons/widgets/space_widgets.dart';
+import 'package:aplikasi_point_of_sale/src/core/commons/widgets/space_widget.dart';
 import 'package:flutter/material.dart';
 
 class CustomDropdown extends StatelessWidget {
